Made more changes to the Water shader. Quality of life changes
Made QOL changes in prep for IBL PBR

+++ GLProjects/GLApp/GLApp/main.cpp
@@ -62,7 +62,7 @@ int main()
 		Tick();
 		Input::DoMovement(deltaTime, MainRenderer->MainCamera, Keys, keyboardlockout, MainRenderer->MainPostProcessSetting, InputMap);
 
-		MainRenderer->RenderLoop(glfwGetTime());
+		MainRenderer->RenderLoop(static_cast<float>(glfwGetTime()));
 		
 		
 		//swap screen buffers
